+++ Monster/Assets/Scripts/ScriptableObject/PlayerStatScriptableObject.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlayerScriptableObject", menuName = "ScriptableObject/Player")]
 public class PlayerStatScriptableObject : ScriptableObject
 {
-    public int maxhealth = 100;
-    public int health = 100;
+    public float maxhealth = 100;
+    public float health = 100;
     public float speed = 1f;
 
     public float attackDamage = 10f;
